\name{address_AixesurVienne}
\alias{address_AixesurVienne}
\docType{data}

\title{
Street addresses of Aixe-sur-Vienne
}

\description{
Street addresses of Aixe-sur-Vienne, a town of approximately 5800 inhabitants in Haute-Vienne (France).
}

\usage{address_AixesurVienne}

\format{
A factor containing all 1911 existing street address numbers.
}

\source{From an open platform for French public data:
 
\url{https://www.data.gouv.fr/fr/datasets/base-d-adresses-nationale-ouverte-bano/} (\url{http://bano.openstreetmap.fr/data/}).
}

