\name{address_Limoges}
\alias{address_Limoges}
\docType{data}

\title{
Street addresses of Limoges
}

\description{
Street addresses of Limoges, a city of approximately 133600 inhabitants in Haute-Vienne (France).
}

\usage{address_Limoges}

\format{
A factor containing all 35975 existing street address numbers.
}

\source{From an open platform for French public data: 

\url{https://www.data.gouv.fr/fr/datasets/base-d-adresses-nationale-ouverte-bano/} (\url{http://bano.openstreetmap.fr/data/}).
}
