\name{address_PierreBuffiere}
\alias{address_PierreBuffiere}
\docType{data}

\title{
Street addresses of Pierre-Buffiere
}

\description{
Street addresses of Pierre-Buffiere, a small town of approximately 1200 inhabitants in Haute-Vienne (France).
}


\usage{data(address_PierreBuffiere)}

\format{
A factor containing all 346 existing street address numbers.
}


\source{From an open platform for French public data: 

\url{https://www.data.gouv.fr/fr/datasets/base-d-adresses-nationale-ouverte-bano/} (\url{http://bano.openstreetmap.fr/data/}).
}

