\name{census}
\alias{census}
\docType{data}

\title{
Alabama census
}
\description{
Populations in Alabama cities and towns.
}

\usage{data(census)}

\format{
A data frame containing the populations of all 460 Alabama cities or towns (dimension: one row and 460 columns).
}

\source{From the United States Census Bureau:

\url{https://factfinder.census.gov/faces/tableservices/jsf/pages/productview.xhtml?src=bkmk}.
}

